使用 #1770 的insertItemNoPolling方法替换调用栈检查
