+++ src/main/java/dev/dubhe/anvilcraft/block/BlockPlacerBlock.java
@@ -48,7 +48,7 @@ import java.util.EnumSet;
 import java.util.List;
 
 import static dev.dubhe.anvilcraft.api.entity.player.AnvilCraftBlockPlacer.anvilCraftBlockPlacer;
-import static dev.dubhe.anvilcraft.util.ItemHandlerUtil.getSourceItemHandlerRecursive;
+import static dev.dubhe.anvilcraft.api.itemhandler.ItemHandlerUtil.getSourceItemHandlerRecursive;
 
 @ParametersAreNonnullByDefault
 @MethodsReturnNonnullByDefault
