+++ src/main/java/dev/dubhe/anvilcraft/util/ItemHandlerUtil.java
@@ -1,13 +1,10 @@
 package dev.dubhe.anvilcraft.util;
 
-import dev.dubhe.anvilcraft.AnvilCraft;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.vehicle.ContainerEntity;
-import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
-import net.minecraft.world.level.block.Block;
 import net.minecraft.world.phys.AABB;
 import net.neoforged.neoforge.capabilities.Capabilities;
 import net.neoforged.neoforge.items.IItemHandler;
@@ -15,39 +12,7 @@ import org.jetbrains.annotations.Nullable;
 
 import java.util.List;
 
-import static dev.dubhe.anvilcraft.block.BlockPlacerBlock.ORIENTATION;
-
-public class ItemHandlerUtil implements IItemHandler {
-    @Override
-    public int getSlots() {
-        return 0;
-    }
-
-    @Override
-    public ItemStack getStackInSlot(int slot) {
-        return null;
-    }
-
-    @Override
-    public ItemStack insertItem(int slot, ItemStack stack, boolean simulate) {
-        return null;
-    }
-
-    @Override
-    public ItemStack extractItem(int slot, int amount, boolean simulate) {
-        return null;
-    }
-
-    @Override
-    public int getSlotLimit(int slot) {
-        return 0;
-    }
-
-    @Override
-    public boolean isItemValid(int slot, ItemStack stack) {
-        return false;
-    }
-
+public class ItemHandlerUtil {
     @Nullable
     public static IItemHandler getSourceItemHandler(BlockPos inputBlockPos, Direction context, Level level) {
         if (level == null) return null;
@@ -75,22 +40,5 @@ public class ItemHandlerUtil implements IItemHandler {
             );
         }
         return input;
-    }
-
-    @Nullable
-    public static IItemHandler getSourceItemHandlerRecursive(Block source, BlockPos inputBlockPos, Direction context, Level level) {
-        int i = 0;
-        do {
-            if (level == null) return null;
-            if (level.getBlockState(inputBlockPos).is(source)
-                && level.getBlockState(inputBlockPos).getValue(ORIENTATION).getDirection() == context
-            ) {
-                i++;
-                inputBlockPos = inputBlockPos.relative(context.getOpposite());
-            } else {
-                return getSourceItemHandler(inputBlockPos, context, level);
-            }
-        } while (i < AnvilCraft.config.blockPlacerRecursiveRetrievalDistanceMax);
-        return null;
     }
 }
